show() is not virtual in QWindow
It's better to know if we're shown using showEvent() which is virtual
than show() which isn't. By reimplementing the non-virtual show we break
all uses going through a QWindow* and not a Qt3DRender::QWindow*.

This is necessary for instance to get Qt3DRender::QWindow compatible
with QWindowContainer (which calls ::QWindow::show()).

Change-Id: I7ae367bbb7d539ba6caf336893d5d8ce53421d5d
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>