Fix race on QAspectManager shutdown
When QAspectManager::quit() is called, the exec() loop finishes. Then
QAspectThread calls shutdown() and deletes the QAspectManager instance.

However, in certain cases, the QAspectManager object can be deleted before
QAspectThread::quit() has a chance to finish running, causing a crash when
it tries to acquire the m_waitForEndOfExec semaphore.

Make exec() wait for quit() to prevent this.

Change-Id: I579e3e6b5684984ec32497852d94cc050cccbef5
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>