qgltf: Output a KHR_materials_common block
Unless disabled by a command-line switch.

The spec is insufficient to support normal maps, the "normalmap"
value is our own extension to the extension for now.

Also, the common technique name is always PHONG. The correct material
can be chosen based on the presence and type of the value list entries.

Task-number: QTBUG-49384
Change-Id: I4f2e5848ad8019c44e9392416501655ab1163d7c
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>