Fix RenderViewJob crash on Renderer shutdown
Set Renderer::m_running to zero to prevent crashes when Renderer::shutdown
is called while a RenderViewJob is running.

This only affects QtQuick/Scene3D. It's not a problem in the case of pure
Qt3D, as Renderer::m_running is set to zero when Renderer::setSurface(0)
is called before Renderer::shutdown.

Change-Id: I41f8da7589c098c6df9d24792932a293df7982aa
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>