Fix resource leaks in ArrayAllocatingPolicy.
Previously we ended up with default constructed elements in the bucket,
and were constructing new ones for every allocation. As some of the types
used in here allocate resources in their ctors themselves we were leaking
those resources.

Now we only construct each element at most once, and we don't construct
any element that is never used. The resulting re-use of elements exposed
the missing cleanup method in tst_ArrayResource.

Change-Id: Ie875fea3ebc9b6f96d0239a6064474fbea798814
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>