QInputAspect: create EventSourceSetterHelper in main thread
Turns out that on initialize is called in the AspectThread.
Since event filters objects need to be created in the same thread
as the event source (the main thread), code has to be corrected.
Now the eventSourceHelper is created directly by the InputHandler when it is
created (as it's created in the main thread). Then onInitialize only takes
care of setting the eventFilterService which isn't available before.

Change-Id: I2cb18372337793230625f699a2ed51c7331af971
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>