Added QBackendNodes for QRenderStates.
RenderStateNode is the backend node for QRenderStates. The actual
state-dependent behavior is implemented in subclasses of
RenderStateImpl. Backend nodes for QRenderPass and QStateSet now hold
IDs of the RenderState-nodes that they contain, instead of pointers.
This means that every frame the RenderView has to get the actual
backend nodes while building the RenderStateSets. This needs to be
fixed.

Change-Id: I908ce9763fdb9f4e851df8dbcecf5dc6ce5a4f49
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>