Force OpenGL texture reload on TextureImage update
After a TextureImage update, the corresponding backend texture node
DNA is not updated until its OpenGL texture is reloaded. This may not
happen if the previous texture DNA is the same as the one associated
with a texture unit.

Force a OpenGL texture/DNA update after we have new data from the load
texture job.

Change-Id: Id79ac9c177bb763d986354bdcd0e2736835cb94a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>