render/frontend: eradicate Q_FOREACH loops [low-risk]
... by replacing them with C++11 range-for loops. To
avoid detaches of these mutable Qt containers, wrap
the container in qAsConst(), where needed.

This is the batch with low-risk changes. They operate
on local containers or the loop body clearly does not
cause the container to change.

Saves another ~2.2KiB in text size on optimized GCC 6.0
Linux AMD64 builds.

Change-Id: I4daec2939ab778565c20140244cdf58e88bb8129
Reviewed-by: Volker Krause <volker.krause@kdab.com>