Remove quadratic complexity in ~QEntitiy
The old code Q_FOREACH'ed over d->m_components, taking a
local copy to do so, and calling removeComponent() in the
loop body. But removeComponent() erases elements from
m_components. In doing so, it forces a detach of the
container and hits the worse-case complexity of erase:
always removing the first element, which has quadratic
complexity.

Fix by moving the contents into a local variable before
entering the loop. The removeComponent() will still try
to removeOne() from m_components, but it's now a no-op,
because the container is already empty.

Change-Id: I732c4bdd33bb4778ac7902a8786a186c25b13666
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>