Add QNodeCreatedChange
This is the first step in replacing the cloning used in Qt 3D with
a much simpler system. The basic idea is that now we will send change
events to the backend with enough information for the backend objects
to be initialized. The frontend objects will be responsible for creating
this data and wrapping it in a change event. The change event will be
delivered to the backend, where the contained data can be unwrapped
and used.

For shutdown, a simpler, but similar approach will be used. When a
QNode is deleted, all the backend needs to know is its type and QNodeId.

Task-number: QTBUG-51835
Change-Id: I36bca48ed49eac7f4eae722e5cf5e79663d75bf7
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>