Solve memory corruption in QRenderSurfaceSelector
When using reinterpret_cast<QObject *> on a QSurface pointer we corrupt
things because both QWindow and QOffscreenSurface use multiple
inheritance from QObject followed by QSurface. This led to the QSurface
pointer being interpreted as a QObject pointer but because the QSurace
part of the window or offscreen surface is offset (due to the multiple
inheritance) it led to problems later on. Namely when trying to see if
QVariant was able to convert the QObject surface pointer to a QNodeId.
This in turn led ot undefined behavior - in my case calling the
destructor of the QQuickView.

We need to be much more careful with QSurface casting. Again, another
reason to just make QSurface a QObject in Qt 6.

Change-Id: I515d0f406ab509cecfda75222df5a74d0ed06e1a
Reviewed-by: Volker Krause <volker.krause@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>