Remove todo about making pure virtual
There are some classes that do not need to store any additional data,
such as FrustumCulling framegraph node. So no need to force them to
override with an empty. Instead just have a do nothing implementation
that only logs to the Nodes category so users can use it as a diagnostic
when developing new aspects.

Change-Id: If6a12b7067beafdbaa9de658c96a4746f3209f3b
Reviewed-by: Volker Krause <volker.krause@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>