Do enabled state and parent/child relationship init in framegraph base
Move this out of the creation functor as the peerId is not set at this
point which breaks the setParentId() function. Move this to the
initializeFromPeer() function and have each FrameGraphNode type call
the base class implementation to set the basic stuff up.

With this change, the main loop now runs through using the QT3D_NO_CLONE
env var. However, it does not render again, yet.

Change-Id: Ied12deb4732b5c103e647cc4cfea935114ea31ca
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>