QBlendEquationArguments creates creation changes
We should change the QBlendEquationSeparate class so that it does not
inherit from this one. That would then allow us to simplify this class
to not contain the alpha arguments. Alternatively, we could investigate
folding the separate args version into this by introducing a NotSet
value that we default the alpha arguments to. If they are NotSet then
we use the non-separate blending functions on the backend. If they are
set then we use the separate Rgb + Alpha blending functionality.

Change-Id: I14993b338228d120c6309536b0a286bed9ed870c
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>