Only case to QScenePropertyChange if that's actually what it is
Now that we also have QNodeCreatedChange/QNodeDestroyedChange we need
to protect against casting to incorrect types.

It may well make sense to have a helper on QBackendNode that calls a
specific virtual for the different change types. This would be
analogous to QObject::event() calling e.g. resizeEvent() with an
argument properly cast to the right type.

Change-Id: I5313ba43c39b6e47f3fdf1f8f26f3e6d57a5677c
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>