Introduce a new base class for node property change events
This will allow QScenePropertyChange to be a default implementation
that uses QVariant as the payload for automatic uses. It will also
allow to have a QNodeTypedPropertyChange template subclass for
strongly typed payloads - similar to the creation changes. This will
be useful for common change types like transformations sending a
QMatrix4x4 to avoid QVariant and conversion overheads.

Change-Id: Ia64dcb892b42e2cb2a18a55dfa7f4073ac58b2e3
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>