Scene3DItem: add the hoverEnabled property
If set to true, the item will receive HoverEvents that will be transmitted as
positionChanged events to MouseHandlers in the scene. For most cases this would
result in high traffic for events that aren't handled. If hoverEnabled is set
to true, then that processing will take place, otherwise it won't (unless you
are pressing a mouse button). The default is hoverEnabled set to false.

Change-Id: I8fc2d519ee902349ace8f4e130cd12f3018c01af
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>