PickBoundingVolumeJob: optimize EntityGatherer
Instead of creating lots of small QVectors, use one, moving
it in and out of gatherEntities().

Likewise, in EntityGatherer::entities(), reuse m_entities'
capacity by clearing it and passing it as the initial
vector to the first gatherEntities() call.

As a result, we have reduced the number of allocations from
O(N) to O(log N) for the first entities() call and from O(N)
to (usually, eventually) zero for subsequent calls.

Costs 200B in text size. I tried also passing the vector as
an out parameter, but that increases text size by 600B, so
I chose this approach.

Change-Id: I304a67276429a123b9cc3e2a1a4921c1bfd0d8fa
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>