QNodeCreation overhaul
When a node has its parent set from NULL to some valid parent
it sends a QNodeCreatedChange. The new parent also subsequently receives a
QNodeAddedChange telling him that a child was added to its children property.

When a node with a valid parent changes parent to another valid parent, it
sends a QNodeRemovedChange to the old parent and a QNodeAddedChange to the new parent.

When a node with a valid parent has its parent set to NULL, it first sends a
QNodeRemovedChange to the old parent followed by a QNodeDestroyed change.

When a node is destroyed, prior to destruction it sets its parent to NULL,
which sends notifications as explained above.

Change-Id: Id99727542367797577c2bbb108580dfca902e776
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>