Set QObject parent immediately and defer backend notifications
This is needed to ensure that subtrees created up front before the
aspect engine starts all get added to the scene. This restores correct
behavior in cylinder-cpp.

With this change all QNode data is correctly set up synchronously
upon construction or via setParent(). That is, the scene and arbiter
are properly propagated from the parent and the scene is made
aware of mappings.

Notfications to the backend are deferred until the QNode is fully
constructed.

Unit tests updated to match this and documentation expanded to
capture the details.

Change-Id: Ic6b3b46240fdccd67f9526fc3219aac8b7f350f3
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>