GLTFIO: remove {QMap,QMultiHash}::contains() calls
They are all followed by value() or op[] calls with the same
argument, causing it to be looked up twice each time (once in
contains() and then again in value()).

Just call find() or, for QMultiHash, equal_range(), and check
the iterators to see whether the value was found.

Sprinkle with Q_UNLIKELY as a drive-by (when the consequence
is a qWarning(), which indicates an unlikely event).

Even saves ~1.9KiB in text size on optimized GCC 6.0 Linux
AMD64 builds.

Change-Id: I9bbab8c2ec5eb9070917c60c303b98564fcbd674
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>