GLTFIO: remove QJsonObject::contains() calls
They are all followed by QJsonObject::value() calls with the
same ID, causing the ID to be looked up twice each time (once
in contains() and then again in value()).

Just call value() and check the result for isUndefined() to see
whether the value was found. In some cases, not even that is
necessary, and we just continue to convert the value to, e.g.,
an array.

Sprinkle with Q_UNLIKELY as a drive-by (when the consequence is
a qWarning(), which indicates an unlikely event).

Even saves ~420B in text size on optimized GCC 6.0 Linux AMD64
builds.

Change-Id: If5ead96df3083a0ceeede25ad8f6477c5c092a04
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>