GLTFIO: don't iterate over QJsonObject::keys()
... iterate using (const_)iterators instead.

This avoids the creation of a temporary QList to hold the
keys as well as the following binary search inside value()
on each iteration.

Given the sorry implementation of QJsonObject iterators
(basically just an int, indexing ino QJsonObject::at(int)),
I was surprised to see that this even saves ~1.8KiB in text
size on optimized GCC 6.0 Linux AMD64 builds.

Change-Id: I98c27feeaf26e03fefe67692ea6aeee30d95268c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>