Bring more type safety to the buttons properties
The API uses QList<int> now to benefit from the default support for this
type in QML. We internally store it as a QVector<int> though so that the
whole backend uses QVector<int> for them. It means we still have a
conversion going on but it's not really on any hot patch and those lists
will be small anyway.

This conversion could be eliminated by hacking qv4sequenceobject.cpp in
qtdeclarative but we're close enough to deadline that it's not worth the
effort IMHO.

Change-Id: I7ef16ec351cc54cc418a8f5c340397f58422add3
Task-number: QTBUG-51467
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>