GLTFIO: wrap remaining qWarning()s in Q_UNLIKELY
If, after checking a condition, we issue a qWarning() or similar,
by definition that check is unlikely to be true.

Tell the compiler so it can move the error handling code out of
the normal code path to increase the effective icache size.

Moved conditional code around where possible so that we could
always use Q_UNLIKELY, instead of having to revert to Q_LIKELY
here and there.

Costs ~100B in text size.

Change-Id: I38389d13ff19762db9e2e91a363e7210bdb43aea
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>