Fix for loading using a Mesh
Added code path avoiding cloning for setting the geometry on the QMesh
which is a QGeopmetryRenderer. This is actually much easier without
cloning. We simply create the QGeometry on the backend, move it to the
main thread, then send it over in a change. In the frontend
QGeometryRenderer, we just extract the pointer to the QGeometry and set
it on the geometry property.

Change-Id: I06db5cba52d783323141a7b1d72b326e319b4960
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>