Introduce local macros for declaring types as movable/primitive/shared
The QtCore macros aren't really suited to a namespaced library
like Qt3D. In particular, the use of Q_DECLARE_SHARED is wrong,
because it places the free swap() function in the Qt's namespace
when it should be in the same (inner) namespace the type is in.

Q_DECLARE_TYPEINFO, otoh, is just inconvenient to use because it
requires exiting and later re-entering of the Qt3D namespace(s).

The new macros do this implicitly.

Change-Id: Ic91f447656f08ed59c118ce09cb5ac90ad212955
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>