Add QNodeDynamicPropertyChange class
To be used when notifying about changes to dynamic properties. It
differs from QNodePropertyChangeBase and subclasses in that it uses
the less efficient, but safe, approach of storing the property name
as a QByteArray. This is not needed for static properties as there
the const char* used for the property name is also static data.

Task-number: QTBUG-51494
Change-Id: Ie4e73d5db03cc733b5fb3bf3d584093d88d12b50
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>