ShaderParameterPack: fix parameter passing (II)
BlockToUBO is not a Trivial Type, and not small,
but it benefits from moves (by way of containing
a QHash member) and setUniformBuffer() is a pure
sink for the argument (unconditionally stores it).

The only callers can arrange for an rvalue being
passed as an argument, so the preferred way of
passing BlockToUBO to setUniformBuffer() is by
rvalue reference. We pass by value, because that
is more general and avoids the need for a 2nd
overload. It also expands to less executable
code than just the rvalue reference overload
alone.

Saves 640B(!) in text size on optimized GCC 6.0
Linux AMD64 builds.

Change-Id: I494e9ffe7d651816fb6a3720f1fd6b7d075eb75c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>