Rename NodeAdded/NodeRemoved ChangeFlags
These are not just for adding/removing QNodes. They are more general
in purpose and are used for adding/removing values to/from collection
based properties e.g. QVector<MyValue>. The addition/removal of QNode
pointers (which we translate to QNodeIds) is just the most common case.

If Q_PROPERTY/QObject gains support for properties of collections and
the typical operations on them, then we will be able to catch and
handle most uses of these change types in QNodePrivate just like we do
for setting a Q_PROPERTY.

Task-number: QTBUG-51494
Change-Id: I707de1647554a61accf81dc79bfe58b289dbff7b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>