Qt3D: replace QStringLiteral with QLatin1String when appending or comparing
It makes little sense to use QStringLiteral in the following
situations:

- for comparison to a QString, because operator== is overloaded
  for QLatin1String.
- when constructing an empty QString, because QLatin1String("")
  resolves to sharedEmpty() when converted to QString, which is
  preferable to an independent copy of sharedEmpty().
- for strings which are immediately appended to, or which are
  appended to other strings.

because no dynamic memory allocation is saved by doing so. But if
the only advantage of QStringLiteral does not apply, all its
disadvantages dominate, to wit: injection of calls to the qstring
dtor, non-sharability of data between C strings and QStringLiterals
and among QStringLiterals, and doubled storage requirements.

Fix by replacing QStringLiteral with QLatin1String or QLatin1Char,
as needed.

Ported one use of QString::number(i) to QLatin1Char('0' + i) where
i is guaranteed to be < 10.

Saves 179B and 2150B, resp., in Input and Render text size on
optimized Linux AMD64 GCC 6.0 builds.

Change-Id: I2e8f43ed085875ce387ffddb18bba37ff21f6cf0
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>