StringToInt: add an overload taking QLatin1String
About 50% of callers pass static string literals,
which can be converted from QStringLiteral to
QLatin1String if by adding the overload.

Because of the complicated logic in the existing
overload, no attempt is made in this commit to
hand down the QLatin1String until the actual
comparison. Instead, the QString overload is
called, for now.

Saves almost 1KiB in text size in Render on
optimized GCC 6.0 Linux AMD64 builds.

Change-Id: I37ed154ec49349f251b6a4515c20eb7cb5c89181
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>