Handle renderer shutdown request and assert in render aspect dtor
If the render aspect is deleted without having been unregistered by
the aspect engine, something has gone wrong. Assert on this in the dtor.

Renderer::shutdown() just requests the render thread to exit. When using
a threaded renderer we also have to wake up the render thread by
releasing the renderview semaphore which it may be waiting to acquire
in Renderer::isReadyToSubmit().

As shutdown() only requests renderer shutdown, we introduce a function
specifically for performing OpenGL (or other graphics API) resource
freeing. This is called when exiting the render thread in the case
of threaded rendering, and directly in the case of synchronous
rendering.

Task-number: QTBUG-50044
Task-number: QTBUG-51035
Change-Id: Ie3a5b2aa9f64c7aeebecedd6a56c29816532f126
Reviewed-by: Janne Koskinen <janne.p.koskinen@theqtcompany.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>