assimp plugin: replace QMap with a C array
The keys are statically known, short, and few, so replace the
QMap<QByteArray, OpenMode> with a C array of a {char[2]; OpenMode}
struct, which occupies just 40B in read-only memory. We don't even
need to sort the list, as binary search in a list of ten entries is
slower than linear search (five comparisons each, on average, with the
linear scan having a predictable memory access pattern.

This is a port of dd1d8f51e67ec93031fdd7f7930d63761d1238e6 to the
other copy of this code, which shows that arguments like 'this is just
a tool, not a library' are ill-advised. Code _will_ be copied.

Change-Id: Ie04137f94e487ce998ec077daf655b09bcbbcfc7
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>