qgltf: replace QHash with a C array
The keys are statically known, short, and few, so
replace the QHash<QByteArray, OpenMode> with a C
array of a {char[2]; OpenMode} struct, which
occupies just 40B in read-only memory. We don't
even need to sort the list, as binary search
in a list of ten entries is slower than linear
search (five comparisons each, on average, with
the linear scan having a predictable memory
access pattern.

Saves ~2.5KiB in text size on optimized GCC 6.0
Linux AMD64 builds.

Change-Id: I8ac680dc3a5c3489438d7dd9077aac46cd5708ab
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>