Pass the static meta object to the creation changes
Unfortunately we can't rely upon QObject::metaObject() because the
dynamic metaobjects created by QML may well get destroyed before the
Qt 3D backend has had time to use the metaobjects to lookup the node
mappers for deleting the backend nodes.

QMetaObject at this time has no way of determining if it is a static
or dynamic metaobject so we have to take the same approach as GammaRay
and walk up the QMetaObject::superData() chain to find the most
"derived" metaobject without a dynamic metaobject above it.

Task-number: QTBUG-42353
Change-Id: Ifb8512c75241a7a2d8035dd7384a589e1b9844d9
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>