Fix shader caching
Introduces a specialized ShaderCache class. This is populated with
shaders on demand from render command processing. Likewise for adding
references to shared QOpenGLShaderPrograms from other Shader backend
nodes.

When a Shader backend node is destroyed, it will remove its reference
from the ShaderCache. When a shared QOpenGLShaderProgram has no
remaining references the cache marks it as pending deletion. If it
still has no references the next time the OpenGL submission thread
calls the purge() function on the cache, it will be removed.

With this approach the actual QOpenGLShaderProgram objects are
managed purely by the OpenGL submission thread and are owned by the
cache once added to it.

Task-number: QTBUG-53669
Change-Id: Ia8459b0fdb775148c83202a801ac41cbdd447bcc
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>