Mutex protect the RenderStateCollection
Without the mutex, this class will race when a framegraph contains
multiple branches that result in renderview jobs being created. This
is because one thread will call RenderStateCollection::renderStates()
as part of building render commands. Then some time later (or at the
same time) another thread could call this function too resulting in
undefined behavior because the data changed by the first thread will
not be pushed to the cpu cache of the second thread (or may be only
partially updated) breaking the invariants of this class.

Added a mutex to avoid the data race. In the future we should instead
have the jobs maintain a local cache of such data so that we do not
end up locking the QBackendNodes a lot in the jobs as this will kill
parallelism just as QVariant does. Better to have the backend nodes as
purely read only during job execution even if this means having an
earlier job that populates an intermediate cache like we planned for
the layer filtering (analogous to a relational databases temporary
table concept).

Task-number: QTBUG-53783
Change-Id: I33043dfd23d8b2e01369bdae25b933e3cfec78c8
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>