JobStats write fix
The write of the frame needs to be performed before we start the new frame.
Otherwise, doing the write just after the submission could clear the cache
when jobs (such as texture loading) taking longer than submission are still
running. In turn that results in the jobs never making it into the trace file.

Change-Id: I7f4883fc33a778edd5d266181c5dbff5a8e568a1
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>