Further job improvements
The RenderViewInitialization job used to have to wait for the world transform
job to be completed because it was used to set the viewProjectionMatrix.

This patch adjust dependencies and adds a new job that sets the
viewProjectionMatrix once the transform update job is complete.

This allows to start most of the processing earlier and better use the
available cpu cores.

Change-Id: I77d751e0b20639934cc00be0cd9888f41f684f89
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>