Fix reset in input event filters in shutdown
The event filters can cause reset if they remain registered while the
scene is being shutdown. If the filters are not unregisteded, they
will continue receiving events when the shutdown deallocates them.

Unregister filters when the scene is being shutdown, except the
eventSource filters, since the eventSource might be already deallocated
by that point. Add connection to eventSource destroyed signal to prevent
this.

Task-number: QTBUG-53834
Change-Id: I1dd1875b98021592580dd59a178ad18308c7d67c
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>