QTransform: remove parameter from matrixChanged
In order to reduce the overhead cost of composing the matrix when the
rotation, scale and/or translation are changed, it is only performed when the
getter matrix() is called.

That means that emitting a signal matrixChanged(QMatrix4x4) would force
the composition when setting the scale, rotation or translation which defeats
the purpose.

The parameter was therefore removed from the matrixChanged signal so that it
can be emitted whenever scale, rotation, translation are changed without the
overhead cost of computing the matrix.

Change-Id: I2e69fb24b4ac1dadc4af3bc5474467e4853a83ed
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>