Avoiding badly quantized axis input values
When moving rapidly Qt may generate multiple mouse events per frame.
The old logic was to overwrite the axis values calculated from each
mouse event. It is more correct to accumulate the axis values from all
events in a frame, which is exactly what this patch does. With this
commit, the camera controllers are now much more smooth and feel more
natural to control.

Change-Id: I50c25609702e5f16e303abf8fb9e7697838f7e2e
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>