Make QHandle tests pass on big endian systems
QHandle uses an union internally so the value of handle is composed of
values of three other integer fields, as defined in Data structure.

The value is different on big endian and little endian systems. This
commit introduces a new GET_EXPECTED_HANDLE macro which determines what
the handle should be equal to based on values of those integer fields
and the system endianness. Note that the values of those fields are
verified too in the tests.

Change-Id: I1f1e37a7469995879ff94e8fc2d6b974c1d4359b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>