Quick3DBuffer: do not always convert from QJSValue
In most cases when Quick3DBuffer::setBufferData is called, the QVariant
parameter contains a QJSValue. In some cases though, such as when the data
property of a Quick3DBuffer is set from a C++ data model, the QVariant may
contain a QByteArray directly. Doing the QJSValue conversion in that case
would result in a failure as the data set would be 0. Perform the proper
QVariant conversion based on the contained type.

Change-Id: I263d275fd780f57a67f11c9ca47c5d6f8ddf7a23
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>