Do not create OpenGL context until render aspect is registered
This avoids a race where the renderer creates a context based on the
default format when the render aspect is created. The alternative
fix would have been to simply create the QRenderAspect after
setting the default surface format. However, that would mean being
careful to do that everywhere in the future too. This fixes it
more neatly and completely by creating the renderer in the
render aspect's onRegistered() function.

Whilst in there, this also solves the previous asymmetry in the
onRegistered() and onUnregistred() functions. We now also
unregister backend types and re-register them in onRegistered.
This is necessary because the mappers depend upon the renderer.

Also added a todo to make the unregisterBackendType() functions
public for 5.8.

Task-number: QTBUG-54370
Task-number: QTBUG-53880
Change-Id: I09a774739df069d3210ed0efe6dd504d968b05f3
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>