Qt3DQuickWindow: subclass QWindow
It used to subclass QQuickWinodw which provides a QQmlIncubationController but
is hardwired to use the scenegraph and its own QOpenGLContext.
This lead to examples not working on Android and EGL based platforms as it is
more strict about multiple contexts being used on the same surface.
Instead make it subclass QWindow (which has no associated QOpenGLContext) and
implement our own QQmlIncubationController.

Change-Id: Ieb78ed517840601c804d8b222aa0c3d0118452bb
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>