Renderer: decouple buffer/texture updates from draw commands
This will make it easier to perform the draw command submission concurrently
with the jobs preparing the next frame.

1) Update Buffers, Textures and load shaders
2) Update/Create VAO and store draw information in RenderCommand
   (part of this will later be moved into the RenderCommand builder jobs)
3) Submit the draw commands

Later on it should be able to proceed to the next frame somewhere between 2
and 3. That does imply that everything is cached for the draw commands to be
performed without conflicting if the backend node changes.

Change-Id: Ic54ff85a01621579722035bd194783bc4e61e58d
Task-number: QTBUG-54423
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>