Texture backend refactoring: Add Texture[Image]DataManager
The TextureDataManager was used to associate QTextureImageDataGenerators
with the QTextureImageData instances that they created. This way, no
generator was executed twice.

The same did NOT apply for QTextureGenerators, which were executed for
every single texture that used them, even if they were the same
(as in *generator1.operator==(*generator2) == true).

This modified header provides a generalized template-class that associates
arbitrary generator pointers with data pointers. Each generator->data entry
is referenced by the textures using it. If no more textures are referencing
an entry, it is deleted.

Textures request data items when they are created. The TextureLoaderJobs
then receive a list of generators to be executed by the TextureDataManagers.
When they have executed the generators and created the data items, they
will give them to the TextureDataManager. When building the OpenGL textures,
these data items will be queried.

Change-Id: I63ef9af63f4ced38284db85d4c92e9e596f6c08c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>