Texture backend refactoring: Add GLTexture, GLTextureManager
The backend Texture class had two main purposes:
(1) serve as backend object for QAbstracTextures
(2) maintain a QOpenGLTexture based on the texture data and properties.
    This also includes computing a texture DNA.

These two purposes can (and therefore should) be decoupled very well.
(1) is only about keeping properties and generator pointers up-to-data
with the front-end QAbstractTexture.
(2) is about building and modifying an QOpenGLTexture and uploading
data to the GPU.

A big advantage of de-coupling these two functions into two distinct
classes is that the actual OpenGL textures can be shared between
identical Texture backend objects. So two QAbstractTextures that have
identical properties and generators will be able to share the same
backend OpenGL texture. This eliminates the need for the error-prone
and currently malfunctioning texture DNAs, as the OpenGL texture
objects can be used directly for texture binding.

The GLTexture class takes care of creating and updating an
QOpenGLTexture. Most logic is copied from the old Qt3D::Render::Texture.

The GLTextureManager class is responsible for
* making sure no two identical GLTextures exist
* provides (creates on demand or shares) GLTextures for Texture backend
  nodes.
* keeps a list of unique, non-sharable GLTextures (e.g. FBO attachments)

If a backend texture is modified, it either needs a new GLTexture (if the
GLTexture instance it was using is shared between multiple backend nodes),
or it can just update its GLTexture (if this instance is not shared).
The GLTextureManager makes sure the appropriate action is taken.

By keeping the non-const methods of GLTexture private to the GLTextureManager,
we make sure that the backend nodes do not mess around with shared instances.

Change-Id: I4f8a0216708289d6ac6398ddc87cd714acabf04c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>