Texture backend refactoring
Texture node only tracks frontend properties and a non-owning pointer to the
GLTexture

The backend Texture node basically just keeps track of it's properties,
parameters, and data. It relies on the GLTextureManager to provide
a GLTexture that matches these properties/parameters/data.

If a backend Texture node is created or modified, it is marked dirty
and added to the TextureManager's list of dirty textures. At the start
of each frame, all gathered dirty textures will ask the GLTextureManager
for a GLTexture instance that matches the texture properties.

Change-Id: Ie436e52ce644a489552cad1633f2233b8c15803f
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>